+++ packages/admin-sdk/src/ui/action-button/index.ts
@@ -31,6 +31,16 @@ export type actionButtonAdd = {
     label: string,
 
     /**
+     * Meteor icon of your action button.
+     */
+    meteorIcon?: string,
+
+    /**
+     * Media file types you want the action button to be displayed for.
+     */
+    fileTypes?: string[],
+
+    /**
      * Callback function which will be called once the action button is clicked.
      */
     callback: (entity: string, entityIdList: string[]) => void,