+++ packages/component-library/src/components/layout/mt-empty-state/mt-empty-state.stories.ts
@@ -25,9 +25,17 @@ export default {
       control: "text",
       description: "The text of the link",
     },
+    linkType: {
+      control: { type: "select", options: ["external", "internal"] },
+      description: "The type of the link, determines if it opens in a new tab or not",
+    },
     buttonText: {
       control: "text",
       description: "The text of the action button",
+    },
+    centered: {
+      control: "boolean",
+      description: "Whether the empty state is centered",
     },
   },
 } as Meta;
@@ -38,7 +46,7 @@ const DefaultTemplate: StoryFn = (args) => ({
     return { args };
   },
   template: `
-    <mt-empty-state v-bind="args" />
+      <mt-empty-state v-bind="args" />
   `,
 });
 
@@ -55,7 +63,7 @@ const ExtendedTemplate: StoryFn = (args) => ({
     return { args };
   },
   template: `
-    <mt-empty-state v-bind="args" @button-click="args.onButtonClick" />
+      <mt-empty-state v-bind="args" @button-click="args.onButtonClick" />
   `,
 });