+++ packages/component-library/src/index.ts
@@ -1,5 +1,6 @@
 import MtDatepicker from "./components/form/mt-datepicker/mt-datepicker.vue";
 import MtBanner from "./components/feedback-indicator/mt-banner/mt-banner.vue";
+import MtBadge from "./components/feedback-indicator/mt-badge/mt-badge.vue";
 import MtLoader from "./components/feedback-indicator/mt-loader/mt-loader.vue";
 import MtProgressBar from "./components/feedback-indicator/mt-progress-bar/mt-progress-bar.vue";
 import MtButton from "./components/form/mt-button/mt-button.vue";
@@ -23,6 +24,12 @@ import MtDataTable from "./components/table-and-list/mt-data-table/mt-data-table
 import MtPagination from "./components/table-and-list/mt-pagination/mt-pagination.vue";
 import MtSkeletonBar from "./components/feedback-indicator/mt-skeleton-bar/mt-skeleton-bar.vue";
 import MtToast, { type Toast } from "./components/feedback-indicator/mt-toast/mt-toast.vue";
+import MtSnackbar from "./components/feedback-indicator/mt-snackbar/mt-snackbar.vue";
+import {
+  useSnackbar,
+  type Snackbar,
+} from "./components/feedback-indicator/mt-snackbar/composables/use-snackbar";
+import MtPromoBadge from "./components/feedback-indicator/mt-promo-badge/mt-promo-badge.vue";
 import MtPopover from "./components/overlay/mt-popover/mt-popover.vue";
 import MtPopoverItem from "./components/overlay/mt-popover-item/mt-popover-item.vue";
 import MtPopoverItemResult from "./components/overlay/mt-popover-item-result/mt-popover-item-result.vue";
@@ -57,6 +64,7 @@ import type {
 
 export {
   MtBanner,
+  MtBadge,
   MtLoader,
   MtProgressBar,
   MtButton,
@@ -83,6 +91,8 @@ export {
   MtPagination,
   MtSkeletonBar,
   MtToast,
+  MtSnackbar,
+  MtPromoBadge,
   MtPopover,
   MtPopoverItem,
   MtPopoverItemResult,
@@ -103,6 +113,7 @@ export {
   MtEntitySelect,
   TooltipDirective,
   DeviceHelperPlugin,
+  useSnackbar,
   // @deprecated
   MtBanner as SwBanner,
   // @deprecated
@@ -146,4 +157,4 @@ export {
 };
 
 // Exporting types
-export type { Filter, Option, Toast, ChartOptions };
+export type { Filter, Option, Toast, Snackbar, ChartOptions };