+++ packages/admin-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 6.4.0
+
+### Minor Changes
+
+- 9726930: Add meteorIcon parameter for action button
+- 92c09da: Add save media modal type
+
+## 6.3.0
+
+### Minor Changes
+
+- 2f985e4: feat: Add message type windowRouterGetPath
+- 033631c: Add ability to control smart bar when adding setting item
+
 ## 6.2.1
 
 ### Patch Changes